pyproject.toml
[tool.poetry]
name = "subspace"
version = "0.1.0"
description = ""
authors = ["Dominique Paul <dominique.c.a.paul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
substack-api = "^0.1.0"
pandas = "^2.2.3"
qdrant-client = "^1.12.1"
tqdm = "^4.67.1"
python-dotenv = "^1.0.1"
llama-index = "^0.12.5"
trafilatura = "^2.0.0"
black = "^24.10.0"
cohere = "^5.13.3"
datasets = "^3.2.0"
llama-index-vector-stores-qdrant = "^0.4.0"
llama-index-embeddings-huggingface = "^0.4.0"
llama-index-embeddings-mixedbreadai = "^0.3.0"
llama-index-embeddings-openai = "^0.3.1"
fire = "^0.7.0"
plotly = "^5.24.1"
umap = "^0.1.1"
umap-learn = "^0.5.7"
tiktoken = "^0.8.0"
fastapi = "^0.115.6"
uvicorn = "^0.33.0"
logfire = {extras = ["fastapi"], version = "^2.7.1"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
